setup category hierarchy system to fix #96
